Edge

+++ roles/pgblitz/scripts/main.sh
@@ -20,9 +20,9 @@ echo 'INFO - @Unencrypted PG Blitz Menu' > /var/plexguide/pg.log && bash /opt/pl
 
 ################################################################## CORE
 
-HEIGHT=13
-WIDTH=45
-CHOICE_HEIGHT=6
+HEIGHT=14
+WIDTH=48
+CHOICE_HEIGHT=7
 BACKTITLE="Visit https://PlexGuide.com - Automations Made Simple"
 TITLE="PGDrive /w PG Blitz"
 MENU="Make a Selection:"
@@ -31,7 +31,8 @@ OPTIONS=(A "RClone: Config & Establish"
          B "JSON  : For TeamDrive"
          C "E-Mail: Share Generator for PG Blitz"
          D "Deploy: PG Drive & PG Blitz"
-         E "Tshoot: Remove All Keys (Baseline)"
+         E "Tshoot: Baseline PG Blitz (Fresh Start)"
+         F "Tshoot: Disable PGBlitz"
          Z "Exit")
 
 CHOICE=$(dialog --clear \
@@ -161,15 +162,51 @@ echo 'INFO - Configured RCLONE for PG Drive' > /var/plexguide/pg.log && bash /op
             dialog --title "NOTE" --msgbox "\nPG Drive & PG Blitz Deployed!!" 0 0
             ;;
         E)
+        #### RECALL VARIABLES START
+        tdrive=$(grep "tdrive" /root/.config/rclone/rclone.conf)
+        gdrive=$(grep "gdrive" /root/.config/rclone/rclone.conf)
+        #### RECALL VARIABLES END
+
+        ### Checkers
+        if [ "$gdrive" != "[gdrive]" ]; then
+          echo 'FAILURE - Must Configure gdrive for RCLONE' > /var/plexguide/pg.log && bash /opt/plexguide/roles/log/log.sh
+          dialog --title "WARNING!" --msgbox "\nGDrive for RClone Must be Configured for PG Blitz!\n\nYou Never had GDrive Configured, so cannot baseline!" 0 0
+          bash /opt/plexguide/roles/pgblitz/scripts/main.sh
+          exit
+        fi
+
+        if [ "$tdrive" != "[tdrive]" ]; then
+          echo 'FAILURE - USING ST2: Must Configure tdrive for RCLONE' > /var/plexguide/pg.log && bash /opt/plexguide/roles/log/log.sh
+          dialog --title "WARNING!" --msgbox "\nTDrive for RClone Must be Configured for PG Blitz!\n\nYou Never had TDrive Configured, so cannot baseline!" 0 0
+          bash /opt/plexguide/roles/pgblitz/scripts/main.sh
+          exit
+        fi
+
             dialog --infobox "Baselining PGBlitz (Please Wait)" 3 25
             sleep 2
             systemctl stop pgblitz 1>/dev/null 2>&1
             systemctl disable pgblitz 1>/dev/null 2>&1
+            grep -A10 'tdrive' rclone.conf|grep -v "tdrive" > /root/.config/rclone/tdrive.save
+            grep -A10 'gdrive' rclone.conf|grep -v "gdrive" > /root/.config/rclone/gdrive.save
             rm -r /root/.config/rclone/rclone.conf 1>/dev/null 2>&1
             rm -r /opt/appdata/pgblitz/keys/unprocessed/* 1>/dev/null 2>&1
             rm -r /opt/appdata/pgblitz/keys/processed/* 1>/dev/null 2>&1
             rm -r /opt/appdata/pgblitz/keys/badjson/* 1>/dev/null 2>&1
+
+            echo "[tdrive]" > /root/.config/rclone/rclone.conf
+            cat /root/.config/rclone/tdrive.save > /root/.config/rclone/rclone.conf
+            echo "" > /root/.config/rclone/rclone.conf
+            echo "[gdrive]" > /root/.config/rclone/rclone.conf
+            cat /root/.config/rclone/gdrive.save >> /root/.config/rclone/rclone.conf
+            rm -r /root/.config/rclone/tdrive.save
+            rm -r /root/.config/rclone/gdrive.save
+
             dialog --title "NOTE" --msgbox "\nKeys Cleared!\n\nYou must reconfigure RClone and Repeat the Process Again!" 0 0
+            ;;
+         F)
+            sudo systemctl stop pgblitz 1>/dev/null 2>&1
+            sudo systemctl rm pgblitz 1>/dev/null 2>&1
+            dialog --title "NOTE" --msgbox "\nPG Blitz is Disabled!\n\nYou must rerun PGDrives & PGBlitz to Enable Again!" 0 0
             ;;
         Z)
             exit 0 ;;
